Update index to group specs and show proper titles
The algorithm that determines which spec in a group is the current work
is wrong in a few specs, but I'm not sure if there's a programmatic way
to tell.

This commit also updates the build workflow to use Fonts.html, rather than
Overview.html, as the directory index for css-fonts-3.