Make GH Pages deployment create symlinks rather than redirects
This change makes the GH pages deployment create symlinks for all
un-leveled shortnames, pointing to the appropriate leveled shortnames —
eliminating the need for the web server to do any rewrites, and
eliminating the need for any redirects (either server- or client-side).
Thus, it replaces the existing build mechanism that generated index.html
files containing client-side redirects.

This commit is a port of w3c/csswg-drafts#7886.

Co-authored-by: Michael[tm] Smith <mike@w3.org>