Generate timestamps.json file w/ git timestamps for sources
For the GH Pages deployment, this change generates a timestamps.json file
that maps each shortname in the repo to a timestamp representing the time
for the most recent commit of corresponding Bikeshed source file for that
shortname — enabling any downstream users/consumers of the deployed
GH Pages specs to have accurate last-committed data for all sources.

This commit is a port of w3c/csswg-drafts#7898.

Co-authored-by: Michael[tm] Smith <mike@w3.org>