Do full clone rather than shallow clone
This change causes the GH Pages build to do a full clone rather than a
shallow clone. This is necessary in order to get the correct timestamps
for all sources (see 8a88ad6).

Otherwise, without this change, git doesn’t report accurate timestamps
for any sources; instead it reports the same timestamp for all of them —
which is apparently whatever the current time is when git creates the
(shallow) clone.

This commit is a port of w3c/csswg-drafts#7932.

Co-authored-by: Michael[tm] Smith <mike@w3.org>